moved login button to home fragment for better UI, kept switch user in the dropdown
