+++ Troubadour/app/src/main/res/layout/fragment_home.xml
@@ -20,5 +20,14 @@
         android:layout_centerVertical="true"
         android:layout_centerHorizontal="true" />
 
+    <Button
+        android:id="@+id/loginButton"
+        android:text="Login to Spotify"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/generatePlaylistButton"
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="53dp" />
+
 
 </RelativeLayout>
